Refactor TransformText into a separate component and integrate it with ProfessionButton and VisualTeaser components.
